fix: config timer float type

+++ server/Code/MorpehFeatures/RoomPokerFeature/Components/RoomPokerStats.cs
@@ -9,5 +9,5 @@ public struct RoomPokerStats : IComponent
     public CurrencyType CurrencyType;
     public long Contribution;
     public long BigBet;
-    public int TurnTime;
+    public float TurnTime;
 }
